function print_int

+++ print_int.c
@@ -9,17 +9,16 @@
  */
 int print_int(va_list arg)
 {
-	int num;
+	long int num;
 	int len = 0;
-	int div = 1;
+	long int div = 1;
 
 	num = va_arg(arg, int);
 
-	if (num < 0)
+	if (num < 0 )
 	{
 		len += my_write('-');
 		num *= -1;
-
 	}
 	while ((num / div) > 9)
 		div *= 10;
